launch videoIdent on the main thread instead of a background thread
For me the app would crash because some calls are supposed to run on the main thread, which makes sense for UI.

I think I will need to do a few more PRs. Are you still maintaining this and do want more PRs or should I just fork it and call it a day?